feat: comprehensive bugfix audit + music integration + OTA + hardware simplification
## Critical Bug Fixes (Phase 1)
- AssessmentIndicators: r.results → r.assessments — saved assessments now display correctly
- assessments.py: window_sec uses timestamp column instead of accel_x (was computing nonsense)
- auth/org.py: effective_org_id returns DEFAULT_ORG_ID instead of empty string

## Medium Bug Fixes (Phase 2)
- session.py: music_bpm/music_duration changed from Integer to Float + Alembic migration
- config.py: WiFi config requires auth (GET: teacher+, PUT: admin+)
- sessions.py: delete_session now cleans up AssessmentResult + SessionEvaluation
- client.ts: uploadSessionMusic uses native fetch (avoids FormData/Content-Type conflict)
- api.test.ts: Updated to use getDeviceWifiConfig

## Low Priority Fixes (Phase 3)
- firmware.py: Removed dead POST /api/firmware/ack endpoint + unused imports
- sessions.py: Fixed empty pass in music upload (now assigns session.description)
- assessments.py: Removed dead class_obj query
- admin.py: All endpoints use Pydantic model_validate() for proper serialization
- context.tsx: UserInfo imports from client.ts (deduplication)

## Documentation Fixes (Phase 4)
- Port 8080→8000 in start.sh, docker-compose.yml, Makefile
- FirmwareUpload.tsx: OTA text updated to GitHub Pages + 24h
- hardware/README.md: Removed ME6211 from PCB diagram + test steps
- hardware/schematic.md: Removed SS12 + ME6211 from routing diagram
- README.md: Removed ME6211 from BOM, updated OTA directory tree
- OPERATION.md: Updated OTA section to reflect GitHub Pages workflow

## Frontend fetch → api client (Phase 5)
- 10 pages converted to use centralized api.* methods
- 18 new API methods added to client.ts
- All pages now support API_BASE for production/development switching

## Music Integration Complete
- librosa analysis: BPM, beat_times, downbeat, cadence, intensity, stops
- RealtimeAnalyzer: 1500-frame buffer, 250-frame rhythm cadence
- WebSocket: rhythm_update/freeze_update persisted to AnalysisResult DB
- Frontend: BeatIndicator, LiveView music player, SessionDetail music settings
- Session model: 7 music columns + music_url

## OTA System
- ESP32 checks GitHub Pages static files every 24 hours
- HTTPS with embedded ISRG Root X1 CA
- No backend ack reporting needed

## Hardware Simplification
- ESP32-C3 Super Mini replaces custom MINI-1 + ME6211 + SS12
- Single external 16500 battery only (no charging circuit)
- Battery thresholds: 3200mV low warning, 2800mV cutoff

+++ backend/app/models/session.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime
-from sqlalchemy import Column, String, DateTime, Enum, Integer, JSON, ForeignKey, Text
+from sqlalchemy import Column, String, DateTime, Enum, Integer, JSON, ForeignKey, Text, Float
 from app.db.base import Base
 
 
@@ -27,9 +27,9 @@ class Session(Base):
     start_time = Column(DateTime, nullable=True)
     end_time = Column(DateTime, nullable=True)
     music_file = Column(String(500), nullable=True)
-    music_bpm = Column(Integer, nullable=True)
+    music_bpm = Column(Float, nullable=True)
     music_beat_times = Column(JSON, nullable=True)
     music_stop_times = Column(JSON, nullable=True)
-    music_duration = Column(Integer, nullable=True)
+    music_duration = Column(Float, nullable=True)
     music_element = Column(String(100), nullable=True)
     music_url = Column(String(1000), nullable=True)
